+++ database/bigtable/utils.go
@@ -1,0 +1,11 @@
+package bigtable
+
+//Createbigtable json formation.
+func Createbigtabledictnoaryconvert(option Createbigtable, Createbigtablejsonmap map[string]interface{}) {
+
+	if option.tableId != ""{
+		Createbigtablejsonmap["tableId"] = option.tableId
+	}
+
+	Createbigtablejsonmap["table"] = option.table
+}
