+++ database/dynamodb/dynamodb.go
@@ -1,14 +1,7 @@
 package dynamodb
 
-import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-	awsauth "github.com/cloudlibz/gocloud/awsauth"
-	"io/ioutil"
-	"net/http"
-)
 
+//List tables.
 func (dynamodb *Dynamodb) Listtables(request interface{}) (resp interface{}, err error) {
 
 	param := request.(map[string]interface{})
@@ -47,6 +40,7 @@ func (dynamodb *Dynamodb) Listtables(request interface{}) (resp interface{}, err
 	return resp, err
 }
 
+//Delete tables.
 func (dynamodb *Dynamodb) Deletetables(request interface{}) (resp interface{}, err error) {
 	param := request.(map[string]interface{})
 
@@ -78,6 +72,7 @@ func (dynamodb *Dynamodb) Deletetables(request interface{}) (resp interface{}, e
 	return resp, err
 }
 
+//Create tables.
 func (dynamodb *Dynamodb) Createtables(request interface{}) (resp interface{}, err error) {
 	param := request.(map[string]interface{})
 	var createtable Createtable
@@ -268,12 +263,10 @@ func (dynamodb *Dynamodb) Createtables(request interface{}) (resp interface{}, e
 	createtablejsonmap := map[string]interface{}{}
 
 
-	fmt.Println(createtable.provisionedThroughput)
 
 	preparecreatetablejsonmap(createtablejsonmap,createtable)
 
-	fmt.Println(createtablejsonmap["ProvisionedThroughput"])
-
+	
 	response := make(map[string]interface{})
 	err = dynamodb.PrepareSignatureV4query(params, createtablejsonmap, response)
 	resp = response
@@ -301,211 +294,10 @@ func preparecreatetablejsonmap(createtablejsonmap map[string]interface{}, create
 	preparekeySchemaparams(createtablejsonmap, createtable)
 	prepareAttributeDefinitionsparams(createtablejsonmap, createtable)
 	prepareGlobalSecondaryIndexesparams(createtablejsonmap, createtable)
-	//prepareLocalSecondaryIndexesparams(createtablejsonmap, createtable)
 }
 
-func preparecreatetableStreamSpecificationparams(createtablejsonmap map[string]interface{}, createtable Createtable) {
 
-	if (createtable.streamSpecification != StreamSpecification{}) {
-
-		streamSpecificationv := make(map[string]interface{})
-
-		if createtable.streamSpecification.StreamViewType != "" {
-			streamSpecificationv["StreamViewType"] = createtable.streamSpecification.StreamViewType
-		}
-
-		if createtable.streamSpecification.StreamEnabled {
-			streamSpecificationv["StreamEnabled"] = createtable.streamSpecification.StreamEnabled
-		}
-
-		createtablejsonmap["StreamSpecification"] = streamSpecificationv
-	}
-}
-
-func preparecreatetableSSESpecificationparams(createtablejsonmap map[string]interface{}, createtable Createtable) {
-	if (createtable.sSESpecification != SSESpecification{}) {
-		sSESpecificationv := make(map[string]interface{})
-		sSESpecificationv["Enabled"] = createtable.sSESpecification.Enabled
-		createtablejsonmap["SSESpecification"] = sSESpecificationv
-	}
-}
-
-func preparecreatetableProvisionedThroughputparams(createtablejsonmap map[string]interface{}, createtable Createtable) {
-	if (createtable.provisionedThroughput != ProvisionedThroughput{}) {
-		provisionedThroughputv := make(map[string]interface{})
-		if createtable.provisionedThroughput.ReadCapacityUnits != 0 {
-			provisionedThroughputv["ReadCapacityUnits"] = createtable.provisionedThroughput.ReadCapacityUnits
-		}
-		if createtable.provisionedThroughput.WriteCapacityUnits != 0 {
-			provisionedThroughputv["WriteCapacityUnits"] = createtable.provisionedThroughput.WriteCapacityUnits
-		}
-		createtablejsonmap["ProvisionedThroughput"] = provisionedThroughputv
-	}
-}
-
-func prepareAttributeDefinitionsparams(createtablejsonmap map[string]interface{}, createtable Createtable) {
-	if len(createtable.attributeDefinitions) != 0 {
-		attributeDefinitionvs := make([]map[string]interface{}, 0)
-
-		for i := 0; i < len(createtable.keySchema); i++ {
-			attributeDefinitionv := make(map[string]interface{})
-
-			if createtable.attributeDefinitions[i].AttributeName != "" {
-				attributeDefinitionv["AttributeName"] = createtable.attributeDefinitions[i].AttributeName
-			}
-
-			if createtable.attributeDefinitions[i].AttributeType != "" {
-				attributeDefinitionv["AttributeType"] = createtable.attributeDefinitions[i].AttributeType
-			}
-
-			attributeDefinitionvs = append(attributeDefinitionvs, attributeDefinitionv)
-		}
-
-		createtablejsonmap["AttributeDefinitions"] = attributeDefinitionvs
-	}
-}
-
-func preparekeySchemaparams(createtablejsonmap map[string]interface{}, createtable Createtable) {
-	if len(createtable.keySchema) != 0 {
-		keySchemavs := make([]map[string]interface{}, 0)
-
-		for i := 0; i < len(createtable.keySchema); i++ {
-			keySchemav := make(map[string]interface{})
-
-			if createtable.keySchema[i].AttributeName != "" {
-				keySchemav["AttributeName"] = createtable.keySchema[i].AttributeName
-			}
-
-			if createtable.keySchema[i].KeyType != "" {
-				keySchemav["KeyType"] = createtable.keySchema[i].KeyType
-			}
-
-			keySchemavs = append(keySchemavs, keySchemav)
-		}
-
-		createtablejsonmap["KeySchema"] = keySchemavs
-	}
-}
-
-func prepareGlobalSecondaryIndexesparams(createtablejsonmap map[string]interface{}, createtable Createtable) {
-
-	if len(createtable.globalSecondaryIndexes) != 0 {
-
-		globalSecondaryIndexesvarrayjsonmap := make([]map[string]interface{}, 0)
-
-		for i := 0; i < len(createtable.globalSecondaryIndexes); i++ {
-
-			globalSecondaryIndexessvjsonmap := make(map[string]interface{})
-
-			if createtable.globalSecondaryIndexes[i].IndexName != "" {
-				globalSecondaryIndexessvjsonmap["IndexName"] = createtable.globalSecondaryIndexes[i].IndexName
-			}
-
-			p := Projection{}
-
-			if createtable.globalSecondaryIndexes[i].projection.ProjectionType == p.ProjectionType && len(createtable.globalSecondaryIndexes[i].projection.NonKeyAttributes) == len(p.NonKeyAttributes) {
-
-				projectionv := make(map[string]interface{})
-				projectionv["ProjectionType"] = createtable.globalSecondaryIndexes[i].projection.ProjectionType
-				projectionv["NonKeyAttributes"] = createtable.globalSecondaryIndexes[i].projection.NonKeyAttributes
-				globalSecondaryIndexessvjsonmap["Projection"] = projectionv
-			}
-
-			if (createtable.globalSecondaryIndexes[i].provisionedThroughput != ProvisionedThroughput{}) {
-
-				provisionedThroughputv := make(map[string]interface{})
-				provisionedThroughputv["ReadCapacityUnits"] = createtable.globalSecondaryIndexes[i].provisionedThroughput.ReadCapacityUnits
-				provisionedThroughputv["WriteCapacityUnits"] = createtable.globalSecondaryIndexes[i].provisionedThroughput.WriteCapacityUnits
-				globalSecondaryIndexessvjsonmap["ProvisionedThroughput"] = provisionedThroughputv
-			}
-
-			if len(createtable.globalSecondaryIndexes[i].keySchema) != 0 {
-
-				keySchemavs := make([]map[string]interface{}, 0)
-
-				for i := 0; i < len(createtable.globalSecondaryIndexes[i].keySchema); i++ {
-
-					keySchemav := make(map[string]interface{})
-
-					if createtable.globalSecondaryIndexes[i].keySchema[i].AttributeName != "" {
-						keySchemav["AttributeName"] = createtable.globalSecondaryIndexes[i].keySchema[i].AttributeName
-					}
-
-					if createtable.localSecondaryIndexes[i].keySchema[i].KeyType != "" {
-						keySchemav["KeyType"] = createtable.globalSecondaryIndexes[i].keySchema[i].KeyType
-					}
-
-					keySchemavs = append(keySchemavs, keySchemav)
-				}
-
-				globalSecondaryIndexessvjsonmap["KeySchema"] = keySchemavs
-			}
-
-			globalSecondaryIndexesvarrayjsonmap = append(globalSecondaryIndexesvarrayjsonmap, globalSecondaryIndexessvjsonmap)
-		}
-
-		createtablejsonmap["GlobalSecondaryIndexes"] = globalSecondaryIndexesvarrayjsonmap
-	}
-
-}
-
-func prepareLocalSecondaryIndexesparams(createtablejsonmap map[string]interface{}, createtable Createtable) {
-
-	if len(createtable.localSecondaryIndexes) != 0 {
-
-		localSecondaryIndexesvarrayjsonmap := make([]map[string]interface{}, 0)
-
-		for i := 0; i < len(createtable.localSecondaryIndexes); i++ {
-
-			localSecondaryIndexesvjsonmap := make(map[string]interface{})
-
-			if createtable.localSecondaryIndexes[i].IndexName != "" {
-				localSecondaryIndexesvjsonmap["IndexName"] = createtable.localSecondaryIndexes[i].IndexName
-			}
-
-			p := Projection{}
-
-			if createtable.localSecondaryIndexes[i].projection.ProjectionType == p.ProjectionType && len(createtable.localSecondaryIndexes[i].projection.NonKeyAttributes) == len(p.NonKeyAttributes) {
-				projectionv := make(map[string]interface{})
-				projectionv["ProjectionType"] = createtable.localSecondaryIndexes[i].projection.ProjectionType
-				projectionv["NonKeyAttributes"] = createtable.localSecondaryIndexes[i].projection.NonKeyAttributes
-				localSecondaryIndexesvjsonmap["Projection"] = projectionv
-			}
-
-
-			if len(createtable.localSecondaryIndexes[i].keySchema) != 0 {
-
-				lenv := len(createtable.localSecondaryIndexes[i].keySchema)
-
-				fmt.Println("i :",i)
-				fmt.Println("lenv", lenv)
-
-				keySchemavs := make([]map[string]interface{}, 0)
-
-				for j := 0; j < lenv ; i++ {
-
-					keySchemav := make(map[string]interface{})
-						fmt.Println("j :",j)
-						fmt.Println("createtable.localSecondaryIndexes[i].keySchema[j].AttributeName : ",createtable.localSecondaryIndexes[i].keySchema[j].AttributeName)
-						keySchemav["AttributeName"] = createtable.localSecondaryIndexes[i].keySchema[j].AttributeName
-
-					if createtable.localSecondaryIndexes[i].keySchema[j].KeyType != "" {
-						keySchemav["KeyType"] = createtable.localSecondaryIndexes[i].keySchema[j].KeyType
-					}
-
-					keySchemavs = append(keySchemavs, keySchemav)
-				}
-
-				localSecondaryIndexesvjsonmap["KeySchema"] = keySchemavs
-			}
-
-			localSecondaryIndexesvarrayjsonmap = append(localSecondaryIndexesvarrayjsonmap, localSecondaryIndexesvjsonmap)
-		}
-		createtablejsonmap["LocalSecondaryIndexes"] = localSecondaryIndexesvarrayjsonmap
-	}
-
-}
-
+//Describe tables.
 func (dynamodb *Dynamodb) Describetables(request interface{}) (resp interface{}, err error) {
 
 	param := request.(map[string]interface{})
@@ -536,71 +328,4 @@ func (dynamodb *Dynamodb) Describetables(request interface{}) (resp interface{},
 	err = dynamodb.PrepareSignatureV4query(params, deletetablesjsonmap, response)
 	resp = response
 	return resp, err
-}
-
-func preparedescribetables(params map[string]string, TableName string, Region string) {
-
-	if TableName != "" {
-		params["TableName"] = TableName
-	}
-
-	if Region != "" {
-		params["Region"] = Region
-	}
-
-	params["amztarget"] = "DynamoDB_20120810.DescribeTable"
-}
-
-func preparedeletetables(params map[string]string, TableName string, Region string) {
-
-	if TableName != "" {
-		params["TableName"] = TableName
-	}
-
-	if Region != "" {
-		params["Region"] = Region
-	}
-
-	params["amztarget"] = "DynamoDB_20120810.DeleteTable"
-}
-
-func preparelisttables(params map[string]string, ExclusiveStartTableName string, Limit string, Region string) {
-	if ExclusiveStartTableName != "" {
-		params["ExclusiveStartTableName"] = ExclusiveStartTableName
-	}
-
-	if Limit != "" {
-		params["Limit"] = Limit
-	}
-
-	if Region != "" {
-		params["Region"] = Region
-	}
-	params["amztarget"] = "DynamoDB_20120810.ListTables"
-}
-
-//PrepareSignatureV4query creates PrepareSignatureV4 for request.
-func (dynamodb *Dynamodb) PrepareSignatureV4query(params map[string]string, paramsmap map[string]interface{}, response map[string]interface{}) error {
-	ECSEndpoint := "https://dynamodb." + params["Region"] + ".amazonaws.com"
-	fmt.Println("ECSEndpoint : ", ECSEndpoint)
-	service := "dynamodb"
-	method := "POST"
-	host := service + "." + params["Region"] + ".amazonaws.com"
-	fmt.Println("host : ", host)
-	ContentType := "application/x-amz-json-1.0"
-	signedheaders := "content-type;host;x-amz-date;x-amz-target"
-	amztarget := params["amztarget"]
-
-	requestparametersjson, _ := json.Marshal(paramsmap)
-	requestparametersjsonstring := string(requestparametersjson)
-	requestparametersjsonstringbyte := []byte(requestparametersjsonstring)
-	client := new(http.Client)
-	request, _ := http.NewRequest("POST", ECSEndpoint, bytes.NewBuffer(requestparametersjsonstringbyte))
-	request = awsauth.SignatureV4(request, requestparametersjsonstringbyte, amztarget, method, params["Region"], service, host, ContentType, signedheaders)
-	resp, err := client.Do(request)
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
-	response["body"] = string(body)
-	response["status"] = resp.StatusCode
-	return err
 }
